+++ tests/testthat/test-class_branch.R
@@ -140,6 +140,8 @@ tar_test("branch$produce_record() of a successful branch", {
   local$run()
   meta <- local$meta
   target <- pipeline_get_target(pipeline, target_get_children(map)[2L])
+  target$file$hash <- hash_object(123L)
+  target$file$bytes <- 16
   record <- target_produce_record(target, pipeline, meta)
   expect_silent(record_validate(record))
   expect_true(grepl("^y_", record$name))