+++ tests/testthat/test-class_junction.R
@@ -1,19 +1,42 @@
-tar_test("junction deps", {
+tar_test("junction with deps", {
   x <- junction_init("x", letters, list(a = LETTERS, b = rev(letters)))
   out <- x$deps
   exp <- data_frame(a = LETTERS, b = rev(letters))
   expect_equal(out, exp)
+  expect_true(x$has_deps)
 })
 
-tar_test("junction_get_splits()", {
+tar_test("junction without deps", {
+  skip_cran()
+  x <- junction_init("x", letters, list())
+  expect_equal(x$deps, data.frame())
+  expect_false(x$has_deps)
+})
+
+tar_test("junction_length()", {
   x <- junction_init("x", letters, list(a = LETTERS, b = rev(letters)))
-  expect_equal(junction_get_splits(x), letters)
+  expect_equal(junction_length(x), length(letters))
+})
+
+tar_test("junction_splits()", {
+  x <- junction_init("x", letters, list(a = LETTERS, b = rev(letters)))
+  expect_equal(junction_splits(x), letters)
+})
+
+tar_test("junction_extract_index()", {
+  x <- junction_init("x", letters, list(a = LETTERS, b = rev(letters)))
+  expect_equal(junction_extract_index(x, "j"), 10L)
+})
+
+tar_test("junction_extract_deps()", {
+  x <- junction_init("x", letters, list(a = LETTERS, b = rev(letters)))
+  expect_equal(junction_extract_deps(x, 10L), c("J", "q"))
 })
 
 tar_test("junction_invalidate()", {
   x <- junction_init("x", letters, list(a = LETTERS, b = rev(letters)))
   junction_invalidate(x)
-  expect_equal(junction_get_splits(x), rep(NA_character_, length(x$splits)))
+  expect_equal(junction_splits(x), rep(NA_character_, length(x$index)))
 })
 
 tar_test("junction_upstream_edges()", {
@@ -24,32 +47,6 @@ tar_test("junction_upstream_edges()", {
   out <- junction_upstream_edges(junction)
   exp <- data_frame(from = c(x, y), to = c(names, names))
   expect_equal(out, exp)
-})
-
-tar_test("junction_transpose() without deps", {
-  names <- paste0("child_", seq_len(3))
-  junction <- junction_init("parent", names)
-  out <- junction_transpose(junction)
-  exp <- list(deps = character(0), split = "child_1")
-  expect_equal(out[[1]], exp)
-  exp <- list(deps = character(0), split = "child_2")
-  expect_equal(out[[2]], exp)
-  exp <- list(deps = character(0), split = "child_3")
-  expect_equal(out[[3]], exp)
-})
-
-tar_test("junction transpose() with deps", {
-  names <-  paste0("child_", seq_len(3))
-  x <- paste0("x_", seq_len(3))
-  y <- paste0("y_", seq_len(3))
-  junction <- junction_init("parent", names, list(x, y))
-  out <- junction_transpose(junction)
-  exp <- list(deps = sort(c("x_1", "y_1")), split = "child_1")
-  expect_equal(out[[1]], exp)
-  exp <- list(deps = sort(c("x_2", "y_2")), split = "child_2")
-  expect_equal(out[[2]], exp)
-  exp <- list(deps = sort(c("x_3", "y_3")), split = "child_3")
-  expect_equal(out[[3]], exp)
 })
 
 tar_test("junction_validate()", {