+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 1.8.0.9009 (development)
+# targets 1.8.0.9010 (development)
 
 * Un-break workflows that use `format = "file_fast"` (#1339, @koefoeden).
 * Fix deadlock in `error = "trim"` (#1340, @koefoeden).
@@ -12,6 +12,7 @@
 * Avoid `store_assert_format()` and `store_convert_object()` is `storage` is `"none"`.
 * Add a `list()` method to `tar_repository_cas()` to make it easier and more efficient to specify custom CAS repositories (#1366).
 * Improve speed and reduce memory consumption by avoiding deep copies of inner environments of target definition objects (#1368).
+* Reduce memory consumption by storing buds and branches as lightweight references when `memory` is `"transient"` (#1364).
 
 # targets 1.8.0